requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asn1crypto==1.5.1
astroid==3.3.6
awsebcli==3.21.0
bcrypt==4.2.0
bidict==0.23.1
blessed==1.20.0
blinker==1.8.2
botocore==1.35.25
cbor2==5.6.5
cement==2.10.14
certifi==2024.8.30
certvalidator==0.11.1
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.2
cryptography==44.0.0
dbfread==2.0.7
dill==0.3.9
dnspython==2.6.1
et-xmlfile==1.1.0
eventlet==0.37.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mccabe==0.7.0
minio==7.2.9
numpy==2.1.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
pandas==2.2.3
pathspec==0.10.1
pika==1.3.2
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.9.0
pylint==3.3.2
pymongo==4.8.0
pyOpenSSL==25.0.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
schedule==1.2.2
semantic-version==2.10.0
setuptools==80.0.0
simple-websocket==1.1.0
six==1.16.0
sqlparse==0.5.1
termcolor==2.4.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
wcwidth==0.2.13
webauthn==2.5.0
Werkzeug==3.0.4
wsproto==1.2.0
